fixed issue #398, replaced Fail with exception in TSpecFun::BetaCf
