+++ src/glib/base/base.cpp
@@ -73,7 +73,6 @@ int _daylight = 0;
 #include "wch.cpp"
 #include "xfl.cpp"
 #include "xmath.cpp"
-#include "stat.cpp"
 
 #include "blobbs.cpp"
 #include "pgblob.cpp"
