+++ src/glib/base/xmath.h
@@ -500,3 +500,11 @@ public:
     void SaveTxt(const TStr& ValNm, const TStr& FNm) const {
         TFOut FOut(FNm); SaveStat(ValNm, FOut); }
 };
+
+/////////////////////////////////////////////////
+// Statistics
+class TStatFun {
+public:
+	static void ChiSquare(const TFltV& OutValVX, const TFltV& OutValVY, const TInt& Df,
+			TFlt& Chi2, TFlt& P);
+};
